Add Ansible, Vagrant and molecule resources
The roles directory contains every role necessary for the installation,
under two subdirectories: bootstrap and configure. Bootstrap being the
roles that are necessary for using Ansible on the CoreOS host in the
first place, and configure being general Kubernetes configuration.

The samples directory contains samples that can be used to test that the
cluster is properly working.

The molecule.yml file is used for testing purposes.

The Makefile contains shortcuts for all the common actions on the
project.